EOC-469 Remove clients Maps and refactor joing/leaving rooms
