+++ client/src/modules/cohort/index.jsx
@@ -41,7 +41,7 @@ import Breadcrumbs from 'common/components/Breadcrumbs';
 import { getCurrentUser } from 'modules/user/model/selectors';
 import { AbortPromiseException } from 'common/exceptions/AbortPromiseException';
 import { cohortsRoute, makeAbortablePromise } from 'common/utils/helpers';
-import { joinRoom, leaveRoom } from 'common/model/actions';
+import { joinRoom, leaveRoom } from 'sockets';
 import history from 'common/utils/history';
 
 import './Cohort.scss';
@@ -66,7 +66,6 @@ class Cohort extends PureComponent {
 
   componentDidMount() {
     const {
-      currentUser: { id: userId },
       match: {
         params: { id: cohortId }
       }
@@ -78,8 +77,7 @@ class Cohort extends PureComponent {
     const roomConfig = {
       subscribeMetaData: !isUnavailable,
       resourceId: cohortId,
-      roomPrefix: Routes.COHORT,
-      userId
+      roomPrefix: Routes.COHORT
     };
 
     joinRoom(roomConfig);
@@ -107,8 +105,7 @@ class Cohort extends PureComponent {
     const roomConfig = {
       subscribeMetaData: !isUnavailable,
       resourceId: previousCohortId,
-      roomPrefix: Routes.COHORT,
-      userId
+      roomPrefix: Routes.COHORT
     };
 
     if (hasCohortChanged) {
@@ -141,7 +138,6 @@ class Cohort extends PureComponent {
 
   componentWillUnmount() {
     const {
-      currentUser: { id: userId },
       match: {
         params: { id: cohortId }
       }
@@ -150,8 +146,7 @@ class Cohort extends PureComponent {
     const roomConfig = {
       subscribeMetaData: !isUnavailable,
       resourceId: cohortId,
-      roomPrefix: Routes.COHORT,
-      userId
+      roomPrefix: Routes.COHORT
     };
 
     leaveRoom(roomConfig);
