+++ client/src/modules/dashboard/index.jsx
@@ -20,9 +20,7 @@ import CollectionView from 'common/components/CollectionView';
 import FormDialog from 'common/components/FormDialog';
 import { ColorType, Routes } from 'common/constants/enums';
 import Breadcrumbs from 'common/components/Breadcrumbs';
-import { IntlPropType, UserPropType } from 'common/constants/propTypes';
-import { getCurrentUser } from 'modules/user/model/selectors';
-import { enterView, leaveView } from 'common/model/actions';
+import { IntlPropType } from 'common/constants/propTypes';
 import './Dashboard.scss';
 
 class Dashboard extends Component {
@@ -35,26 +33,13 @@ class Dashboard extends Component {
   };
 
   componentDidMount() {
-    const {
-      currentUser: { id },
-      fetchListsMetaData
-    } = this.props;
+    const { fetchListsMetaData } = this.props;
 
     this.setState({ pendingForLists: true });
 
     fetchListsMetaData().finally(() =>
       this.setState({ pendingForLists: false })
     );
-
-    enterView(Routes.DASHBOARD, id);
-  }
-
-  componentWillUnmount() {
-    const {
-      currentUser: { id }
-    } = this.props;
-
-    leaveView(Routes.DASHBOARD, id);
   }
 
   handleDialogVisibility = () =>
@@ -190,7 +175,6 @@ class Dashboard extends Component {
 Dashboard.propTypes = {
   archivedLists: PropTypes.objectOf(PropTypes.object),
   cohortLists: PropTypes.objectOf(PropTypes.object),
-  currentUser: UserPropType.isRequired,
   intl: IntlPropType.isRequired,
   privateLists: PropTypes.objectOf(PropTypes.object),
   viewType: PropTypes.string.isRequired,
@@ -204,7 +188,6 @@ Dashboard.propTypes = {
 const mapStateToProps = state => ({
   archivedLists: getArchivedLists(state),
   cohortLists: getCohortsLists(state),
-  currentUser: getCurrentUser(state),
   privateLists: getPrivateLists(state)
 });
